feat: expose `flipLoadedDie`
